backend/requirements.txt
# ============================================
# Smart City Traffic - Python Dependencies
# ============================================
# Updated: Includes Spark, Hadoop, and MLlib dependencies

# ----- Apache Spark & Hadoop -----
pyspark>=3.5.0            # Spark SQL, DataFrame, Streaming
py4j>=0.10.9              # Python-Java bridge for Spark

# ----- Data Processing -----
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # Parquet file support

# ----- Kafka -----
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# ----- Machine Learning (Spark MLlib + Scikit-learn fallback) -----
scikit-learn>=1.3.0       # Fallback ML library
xgboost>=2.0.0
joblib>=1.3.0

# ----- Geospatial -----
h3>=3.7.0                # Uber's hexagonal grid
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.15.0           # Map visualization

# ----- Web API -----
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0    # WebSocket support
python-socketio>=5.10.0
eventlet>=0.34.0         # Async support for Flask

# ----- Visualization -----
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ----- Jupyter -----
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0

# ----- Utilities -----
python-dotenv>=1.0.0
tqdm>=4.66.0             # Progress bars
requests>=2.31.0

# ----- HDFS/Hadoop Python Client (Optional) -----
hdfs>=2.7.0              # WebHDFS client
snakebite-py3>=3.0.5     # HDFS client (Python 3)

# ----- Development -----
black>=23.0.0            # Code formatting
pytest>=7.4.0            # Testing
